+++ tests/integration/alarmhandler/test_alarm_for_degraded_unknown_health_state.py
@@ -9,6 +9,7 @@ from pytest_bdd import given, parsers, scenario, then, when
 from ska_control_model import HealthState
 from tango import DeviceProxy
 
+from tests.integration.conftest import ResponseData, TestContext
 from tests.integration.tmc.conftest import wait_for_event
 
 namespace = os.getenv("KUBE_NAMESPACE")
@@ -44,7 +45,7 @@ def test_tmc_alarm_for_healthstate_unknown_degraded():
         + "alarm when the {device1} {device2} healthState"
     )
 )
-def configure_alarm_healthstate(response_data, device1, device2):
+def configure_alarm_healthstate(response_data: ResponseData, device1: str, device2: str):
     """Alarm is configured for DEGRADED or UNKNOWN healthstate.
 
     :param response_data: fixture for response data
@@ -73,32 +74,34 @@ def configure_alarm_healthstate(response_data, device1, device2):
 # in this case its Dish Masters.
 # when all devices are available this scenario might not be the same.
 @when(parsers.parse("{device1} and {device2} remain in healthState DEGRADED or UNKNOWN"))
-def check_alarms(device1, device2):
+def check_alarms(device1: str, device2: str, test_context: TestContext):
     """Check devices in healthState DEGRADED or UNKNOWN.
 
     :param device1: tango device1 with healthState DEGRADED or UNKNOWN
     :param device2: tango device2 with healthState DEGRADED or UNKNOWN
+    :param test_context: fixture for the instance of TestContext class to store shared data
     """
     tango_device1 = DeviceProxy(device1)
     tango_device2 = DeviceProxy(device2)
-    pytest.device1_result = tango_device1.read_attribute("telescopehealthState").value
+    test_context.device1_result = tango_device1.read_attribute("telescopehealthState").value
     device2_result = tango_device2.read_attribute("healthState").value
     # If the dish is deployed the value will not be DEGRADED
-    assert pytest.device1_result in (HealthState.DEGRADED, HealthState.UNKNOWN)
+    assert test_context.device1_result in (HealthState.DEGRADED, HealthState.UNKNOWN)
     assert device2_result in (HealthState.DEGRADED, HealthState.UNKNOWN)
 
 
 @then("alarm for healthState DEGRADED or UNKNOWN must be raised with UNACKNOWLEDGE state")
-def check_alarm_state(response_data):
+def check_alarm_state(response_data: ResponseData, test_context: TestContext):
     """Check alarm state.
 
     :param response_data: fixture for response data
+    :param test_context: fixture for the instance of TestContext class to store shared data
     """
     alarm_handler = DeviceProxy("alarm/handler/01")
     alarm_tag = response_data.response["alarm_summary"]["tag"]
-    if pytest.device1_result == HealthState.DEGRADED:
+    if test_context.device1_result == HealthState.DEGRADED:
         alarm_tag = tuple([alarm_tag[0]])
-    elif pytest.device1_result == HealthState.UNKNOWN:
+    elif test_context.device1_result == HealthState.UNKNOWN:
         alarm_tag = tuple([alarm_tag[1]])
     assert wait_for_event(
         alarm_handler, "alarmUnacknowledged", alarm_tag, print_event_details=True, timeout=250.0